+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 * Unreleased
+    * Upgrade STM32duino Core from 1.9.0 to 2.0.0.
+        * MemoryBenchmark: Flash usage increases by 2.6kB across the board, but
+          static RAM goes down by 250 bytes.
+        * AutoBenchmark: No change.
+        * Very little change to AceButton code itself.
+    * Upgrade SparkFun SAMD Core from 1.8.1 to 1.8.3.
+        * No change observed in MemoryBenchmark or AutoBenchmark.
 * 1.9 (2021-06-01)
     * Split [LadderButtonCalibrator](examples/LadderButtonCalibrator)
       from [LadderButtons](examples/LadderButtons) for convenience.